:package: add URL Markify script and requirements file
Add the URL Markify script, which had previously been sitting in a random
folder for months. This script generates a nicely formatted website link in
Markdown syntax.

A requirements.txt file has been added with the depedencies required to run
this new script. For convenience, these can be installed with
`pip install -r requirements.txt` globally or in a virtual env.