Generate MD: Rename YAML Frontmatter key for tag to tags
Use plural "tags" key for the YAML Frontmatter of the generated Markdown file.
Note that it was originally "tag" because Simiki parses that instead of
"tags". However, VuePress and other static generators use "tags", which makes
more sense grammatically.

I will fork Simiki and update it to use "tags", as I still want to use Simiki
for personal/internal wikis, but the project has become unmaintained.